+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## [0.8.8] - Unreleased
 ### Added
 * Databricks WorkspaceTree
+* CLI `build` command
 ### Updated
-* Optimized Python Package to on build-in-time
+* Python packages are built just-in-time
 
 ## [0.8.7] - 2025-08-15
 ### Fixed